AI Examples Crawler/requirements.txt
google-generativeai>=0.7.2
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
google-auth>=2.23.0
